refactor: extract 8 terminal functions to commands/terminal.rs (#127 step 8)
Move create_terminal_session, resize_terminal, write_to_terminal,
close_terminal_session, mark_claude_terminal, is_claude_terminal,
set_terminal_cwd_receive, update_terminal_sync_group to terminal.rs.
Re-exported via pub use terminal::* for API compatibility.
